AS-7
AS-7: (State hosting pattern): The [SearchAppBar] compose was moved up to the tree.
